[Refactor] Package Refactoring - 역할에 따른 package refactoring - Timber추가 - 필요없는 meta-data 추가
