+++ app/src/main/java/org/sopt/sample/presentation/home/model/RepoData.kt
@@ -1,4 +1,4 @@
-package org.sopt.sample.model
+package org.sopt.sample.presentation.home.model
 
 data class RepoData(val img:String, val name: String, val des:String, val viewtype:Int) {
 }
