+++ app/src/main/java/org/sopt/sample/presentation/login/LoginActivity.kt
@@ -1,4 +1,4 @@
-package org.sopt.sample
+package org.sopt.sample.presentation.login
 
 import android.app.Activity
 import android.content.Intent
@@ -9,8 +9,10 @@ import androidx.activity.result.ActivityResultLauncher
 import androidx.activity.result.contract.ActivityResultContracts
 import androidx.appcompat.app.AppCompatActivity
 import com.google.android.material.snackbar.Snackbar
-import org.sopt.sample.model.UserData
 import org.sopt.sample.databinding.ActivityLoginBinding
+import org.sopt.sample.presentation.model.UserData
+import org.sopt.sample.presentation.home.HomeActivity
+import org.sopt.sample.presentation.signup.SignUpActivity
 
 class LoginActivity : AppCompatActivity() {
